Fix makefile to use turbo (#1583)
Before this was failing due to the working directory the Makefile was
attached to. This is why we're running a command against the directory
above. This commit also changes the name of the packages to be both
easier to filter and more suitable for the now slightly more complicated
build logs.