Develop network visible for mobile (#1796)
### Make the Network visible to mobile users on the top-right corner of
page
This PR is an implementation of the following DeWork task:
https://app.dework.xyz/fluidity-money/fluidity-app?taskId=15957bd8-0fdd-4fa8-a5b9-1ffa541ed4f2

The following changes were made: 
1. In `dashboard.tsx`, if `{(isTablet || isMobile) &&
showMobileNetworkButton` is true, the `<ChainSelectorButton>` component
is rendered in the navbar.
2. Create a split flag for for the feature. This is declared by `const
showMobileNetworkButton = showExperiment("feature-network-visible");` in
both `MobileModal.txs` and `dashboard.tsx`. `If == true`, which show new
network visibility feature. If `== false`, will render existing design.
By default this is set to false, so if want to render in meantime can
comment out this line and run with `const showMobileNetworkButton =
true;`.
3. Through the use of ternaries, In `MobileModal.tsx`, if
`showMobileNetworkButton==true`, code will render
`<ChainSelectorButton>` in the navbar div, and will not render in the
body content. If `showMobileNetworkButton==false`, code will not render
in navbar div, and will render `<ChainSelectorButton>` in the mobile
body content.

---------

Co-authored-by: alfie <50436981+owfie@users.noreply.github.com>
Co-authored-by: Alex <71931113+af-afk@users.noreply.github.com>
Co-authored-by: user <alex@iometelabs.io>
Co-authored-by: Glorison Lai <98363943+glo12345@users.noreply.github.com>